+++ src/main/java/com/damino/web/admin/market/orderlist/StoreOrderListController.java
@@ -11,12 +11,19 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.damino.web.admin.salesstatus.SalesStatusService;
+import com.damino.web.admin.salesstatus.SalesVO;
+
 @Controller
+@SessionAttributes({"startdate","enddate"})
 public class StoreOrderListController {
 	@Autowired
 	private StoreOrderListService storeOrderListService;
+	@Autowired
+	private SalesStatusService salesStatusService;
 	
 	@RequestMapping("/orderList.smdo")
 	public ModelAndView getOrderListPage(ModelAndView mav, HttpSession session) {
@@ -37,6 +44,21 @@ public class StoreOrderListController {
 		System.out.println("주문 검색 페이지 열기");
 		String storename = (String)session.getAttribute("storename");
 		List<StoreOrderListVO> storeorderlist = storeOrderListService.getStoreOrderList(storename);
+		mav.addObject("storeorderlist", storeorderlist);
+		mav.setViewName("/order/storeOrderSearch");
+		
+		return mav;
+	}
+	
+	
+	@RequestMapping("/orderSearchResult.smdo")
+	public ModelAndView getOrderSearchReuslt(ModelAndView mav, HttpSession session, SalesVO vo) {
+		System.out.println("주문 검색 결과 열기");
+		String storename = (String)session.getAttribute("storename");
+		List<SalesVO> storeorderlist = salesStatusService.getSalesList(vo);
+		
+		mav.addObject("startdate", vo.getStartdate());//세션
+		mav.addObject("enddate", vo.getEnddate());//세션
 		mav.addObject("storeorderlist", storeorderlist);
 		mav.setViewName("/order/storeOrderSearch");
